System.Log.FastLogger.Date

Description

Formatting time is slow. This package provides mechanisms to cache formatted date.

Date cacher

newTimeCache :: TimeFormat -> IO (IO FormattedTime)

Make IO action which get cached formatted local time. Use this to avoid the cost of frequently time formatting by caching an auto updating formatted time, this cache update every 1 second. more detail in Control.AutoUpdate

simpleTimeFormat :: TimeFormat

A simple time cache using format "%d%b%Y:%T %z"

simpleTimeFormat' :: TimeFormat

A simple time cache using format "%d-%b-%Y %T"