System.Log.FastLogger.Internal

Description

The contents of this module can change at any time without warning.

Documentation

type Buffer = Ptr Word8

type BufSize = Int

The type for buffer size of each core.

defaultBufSize :: BufSize

The default buffer size (4,096 bytes).

getBuffer :: BufSize -> IO Buffer

freeBuffer :: Buffer -> IO ()

toBufIOWith :: Buffer -> BufSize -> (Buffer -> Int -> IO ()) -> Builder -> IO ()

type FD = FD

closeFD :: FD -> IO ()

openFileFD :: FilePath -> IO FD

getStderrFD :: IO FD

getStdoutFD :: IO FD

writeRawBufferPtr2FD :: IORef FD -> Ptr Word8 -> Int -> IO Int

invalidFD :: FD

isFDValid :: FD -> Bool

data Builder

Builders denote sequences of bytes. They are Monoids where mempty is the zero-length sequence and mappend is concatenation, which runs in O(1).

data LogStr

Log message builder. Use (<>) to append two LogStr in O(1).

Constructors

LogStr !Int Builder

logStrLength :: LogStr -> Int

Obtaining the length of LogStr.

fromLogStr :: LogStr -> ByteString

Converting LogStr to ByteString.

class ToLogStr msg where

Types that can be converted to a LogStr. Instances for types from the text library use a UTF-8 encoding. Instances for numerical types use a decimal encoding.

Methods

toLogStr :: msg -> LogStr

mempty :: Monoid a => a

Identity of mappend

Examples

Expand

>>> "Hello world" <> mempty
"Hello world"

>>> mempty <> [1, 2, 3]
[1,2,3]

(<>) :: Semigroup a => a -> a -> a infixr 6

An associative operation.

Examples

Expand

>>> [1,2,3] <> [4,5,6]
[1,2,3,4,5,6]

>>> Just [1, 2, 3] <> Just [4, 5, 6]
Just [1,2,3,4,5,6]

>>> putStr "Hello, " <> putStrLn "World!"
Hello, World!

data SingleLogger

A non-scale but time-ordered logger.

newSingleLogger :: BufSize -> IORef FD -> IO SingleLogger

Creating SingleLogger.

data MultiLogger

A scale but non-time-ordered logger.

newMultiLogger :: Int -> BufSize -> IORef FD -> IO MultiLogger

Creating MultiLogger. The first argument is the number of the internal builders.

writeLogStr :: Buffer -> IORef FD -> LogStr -> IO ()

Writting LogStr using a buffer in blocking mode. The size of LogStr must be smaller or equal to the size of buffer.

writeBigLogStr :: IORef FD -> LogStr -> IO ()

Writting LogStr using a temporary buffer.

class Loggers a where

A class for internal loggers.

Methods

stopLoggers :: a -> IO ()

pushLog :: a -> LogStr -> IO ()

flushAllLog :: a -> IO ()

module System.Log.FastLogger.LoggerSet