System.Log.FastLogger.Internal

Description

The contents of this module can change at any time without warning.

Documentation

type Buffer = Ptr Word8

type BufSize = Int

The type for buffer size of each core.

defaultBufSize :: BufSize

The default buffer size (4,096 bytes).

getBuffer :: BufSize -> IO Buffer

freeBuffer :: Buffer -> IO ()

toBufIOWith :: Buffer -> BufSize -> (Buffer -> Int -> IO ()) -> Builder -> IO ()

type FD = FD

closeFD :: FD -> IO ()

openFileFD :: FilePath -> IO FD

getStderrFD :: IO FD

getStdoutFD :: IO FD

writeRawBufferPtr2FD :: IORef FD -> Ptr Word8 -> Int -> IO Int

invalidFD :: FD

isFDValid :: FD -> Bool

data Builder

Builders denote sequences of bytes. They are Monoids where mempty is the zero-length sequence and mappend is concatenation, which runs in O(1).

data LogStr

Log message builder. Use (<>) to append two LogStr in O(1).

Constructors

LogStr !Int Builder

logStrLength :: LogStr -> Int

Obtaining the length of LogStr.

fromLogStr :: LogStr -> ByteString

Converting LogStr to ByteString.

class ToLogStr msg where

Types that can be converted to a LogStr. Instances for types from the text library use a UTF-8 encoding. Instances for numerical types use a decimal encoding.

Methods

toLogStr :: msg -> LogStr

mempty :: Monoid a => a

Identity of mappend

>>> "Hello world" <> mempty
"Hello world"

(<>) :: Semigroup a => a -> a -> a infixr 6

An associative operation.

>>> [1,2,3] <> [4,5,6]
[1,2,3,4,5,6]

newtype Logger

Constructors

Logger (IORef LogStr)

newLogger :: IO Logger

pushLog :: IORef FD -> BufSize -> MVar Buffer -> Logger -> LogStr -> IO ()

flushLog :: IORef FD -> BufSize -> MVar Buffer -> Logger -> IO ()

module System.Log.FastLogger.LoggerSet